pyproject.toml
[project]
name = "print-gateway-server"
version = "1.0.0"
description = "REST API server for centralized print management"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.6",
    "pyyaml>=6.0",
    "pillow>=10.0.0",
]

[project.optional-dependencies]
pdf = ["pypdf>=3.0.0"]
brother-ql = ["brother-ql>=0.9.4"]
cups = ["pycups>=2.0.1"]
all-printers = ["brother-ql>=0.9.4", "pycups>=2.0.1"]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.26.0",
    "ruff>=0.1.0",
]

[project.scripts]
print-server = "src.main:main"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]
